Changed diff download order: align with instructions from TomH in the wiki.

+++ src/bin/fetch_osc.sh
@@ -87,11 +87,11 @@ file_panic()
 
   FILE_PANIC="true"
   until [[ ! -n $FILE_PANIC ]]; do {
-    retry_fetch_file "$REMOTE_PATH/$TDIGIT3.osc.gz" "$LOCAL_PATH/$TDIGIT3.new.osc.gz" "gzip"
+    retry_fetch_file "$REMOTE_PATH/$TDIGIT3.state.txt" "$LOCAL_PATH/$TDIGIT3.new.state.txt" "text"
   }; done
   FILE_PANIC="true"
   until [[ ! -n $FILE_PANIC ]]; do {
-    retry_fetch_file "$REMOTE_PATH/$TDIGIT3.state.txt" "$LOCAL_PATH/$TDIGIT3.new.state.txt" "text"
+    retry_fetch_file "$REMOTE_PATH/$TDIGIT3.osc.gz" "$LOCAL_PATH/$TDIGIT3.new.osc.gz" "gzip"
   }; done
 
   RES_GZIP=`diff -q "$LOCAL_PATH/$TDIGIT3.osc.gz" "$LOCAL_PATH/$TDIGIT3.new.osc.gz"`
@@ -132,25 +132,32 @@ fetch_minute_diff()
   TIMESTAMP=${TIMESTAMP_LINE:10}
 };
 
-if [[ "$REPLICATE_ID" == "auto" && -s "$LOCAL_DIR/state.txt" ]]; then
-  REPLICATE_ID=${$(cat "$LOCAL_DIR/state.txt" | grep -E '^sequenceNumber'):15}
+if [[ -s "$LOCAL_DIR/state.txt" ]]; then
+  MAX_SEQ_NR=$(cat "$LOCAL_DIR/state.txt" | grep -E '^sequenceNumber')
+  MAX_AVAILABLE_REPLICATE_ID=${MAX_SEQ_NR:15}
+  if [[ "$REPLICATE_ID" == "auto" && -s "$LOCAL_DIR/state.txt" ]]; then
+    REPLICATE_ID=$MAX_AVAILABLE_REPLICATE_ID
+  fi
 fi
 
 while [[ true ]];
 do
 {
-  REPLICATE_ID=$(($REPLICATE_ID + 1))
-
-  rm -f "$LOCAL_DIR/state.txt"
-  retry_fetch_file "$SOURCE_DIR/state.txt" "$LOCAL_DIR/state.txt" "text"
-  if [[ -n $FILE_PANIC ]]; then
-    file_panic
+  if [[ $REPLICATE_ID -gt $MAX_AVAILABLE_REPLICATE_ID ]]; then
+    sleep 15
+    rm -f "$LOCAL_DIR/state.txt"
+    retry_fetch_file "$SOURCE_DIR/state.txt" "$LOCAL_DIR/state.txt" "text"
+    if [[ -n $FILE_PANIC ]]; then
+      file_panic
+    fi
+    MAX_SEQ_NR=$(cat "$LOCAL_DIR/state.txt" | grep -E '^sequenceNumber')
+    MAX_AVAILABLE_REPLICATE_ID=${MAX_SEQ_NR:15}
   fi
-  MAX_AVAILABLE_REPLICATE_ID=${$(cat "$LOCAL_DIR/state.txt" | grep -E '^sequenceNumber'):15}
-  echo "$REPLICATE_ID $MAX_AVAILABLE_REPLICATE_ID"
 
-  fetch_minute_diff
-  echo "fetch_osc()@"`date -u "+%F %T"`": new_replicate_diff $REPLICATE_ID $TIMESTAMP" >>$LOCAL_DIR/fetch_osc.log
-  sleep 1
+  if [[ $REPLICATE_ID -le $MAX_AVAILABLE_REPLICATE_ID ]]; then
+    fetch_minute_diff
+    echo "fetch_osc()@"`date -u "+%F %T"`": new_replicate_diff $REPLICATE_ID $TIMESTAMP" >>$LOCAL_DIR/fetch_osc.log
+    REPLICATE_ID=$(($REPLICATE_ID + 1))
+  fi
 };
 done
